+++ app/lib/constants.js
@@ -4,4 +4,5 @@ module.exports = {
   numberOfNearbyResultsToRequest: 4,
   numberOfNearbyResultsToDisplay: 3,
   promHistogramBuckets: [0.01, 0.05, 0.1, 0.2, 0.3, 0.5, 1, 1.5, 5, 10],
+  yourLocation: 'your location',
 };